requirements.txt
alembic==1.7.1
Babel==2.9.1
click==8.0.1
contourpy==1.0.6
cycler==0.11.0
DateTime==4.7
dnspython==2.1.0
email-validator==1.1.3
Faker==9.7.1
Flask==2.2.5
Flask-Babel==2.0.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
fonttools==4.38.0
greenlet==1.1.1
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.4.4
Mako==1.2.2
MarkupSafe==2.0.1
matplotlib==3.5.3
numpy==1.23.5
packaging==21.3
Pillow==9.3.0
psycopg2-binary==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
SQLAlchemy==2.0.17
text-unidecode==1.3
typing-extensions==3.10.0.2
Werkzeug==2.2.3
WTForms==2.3.3
zipp==3.5.0
zope.interface==5.5.2
requests==2.31.0